feat(user): Add refresh sessions api
