+++ src/Module/Api/AuthController.php
@@ -10,6 +10,7 @@ use App\Entity\User;
 use App\Entity\UserSecret;
 use App\Enum\ApiTokenType;
 use App\Enum\ErrorCode;
+use App\Module\Api\Traits\SRPValidationTrait;
 use App\Service\EncryptionService;
 use App\Service\JwtAuthService;
 use Brick\Math\BigInteger;
@@ -37,6 +38,8 @@ use function Windwalker\uid;
 #[Controller]
 class AuthController
 {
+    use SRPValidationTrait;
+
     public function challenge(
         AppContext $app,
         ORM $orm,
@@ -260,19 +263,6 @@ class AuthController
         return compact('accessToken', 'refreshToken');
     }
 
-    public function refreshSessions(\CurrentUser $currentUser, ORM $orm): true
-    {
-        $orm->updateBatch(
-            User::class,
-            [
-                'sess_valid_from' => chronos()
-            ],
-            ['id' => $currentUser->getId()]
-        );
-
-        return true;
-    }
-
     /**
      * @param  \CurrentUser  $currentUser
      *
@@ -296,78 +286,5 @@ class AuthController
         AppContext $app,
     ): true {
         return $app->dispatchController([UserController::class, 'deleteMe']);
-    }
-
-    /**
-     * @param  ORM         $orm
-     * @param  SRPService  $srpService
-     * @param  User        $user
-     * @param  string      $A
-     * @param  string      $M1
-     * @param  string      $sess
-     *
-     * @return  array{ 0: ProofResult, 1: \stdClass, 2: UserSecret }
-     *
-     * @throws NumberFormatException
-     * @throws \Brick\Math\Exception\DivisionByZeroException
-     * @throws \Brick\Math\Exception\MathException
-     * @throws \Brick\Math\Exception\NegativeNumberException
-     */
-    protected function srpValidate(
-        ORM $orm,
-        SRPService $srpService,
-        User $user,
-        string $A,
-        string $M1,
-        string $sess
-    ): array {
-        $userSecret = $orm->mustFindOne(
-            UserSecret::class,
-            ['user_id' => uuid2bin($user->getId())]
-        );
-
-        if (!$loginToken = $user->getLoginToken()) {
-            ErrorCode::INVALID_SESSION->throw();
-        }
-
-        $loginPayload = JWT::decode(
-            $loginToken,
-            new Key($userSecret->getDecodedServerSecret(), 'HS512')
-        );
-
-        $sessPayload = JWT::decode(
-            $sess,
-            new Key($userSecret->getDecodedServerSecret(), 'HS512')
-        );
-
-        if ($loginPayload->sess !== $sessPayload->sess) {
-            ErrorCode::INVALID_SESSION->throw();
-        }
-
-        $password = $user->getPassword();
-
-        $pf = SRPService::decodePasswordVerifier($password);
-
-        $A = BigInteger::fromBase($A, 16);
-        $M1 = BigInteger::fromBase($M1, 16);
-        $b = BigInteger::fromBase($loginPayload->b, 16);
-        $B = BigInteger::fromBase($loginPayload->B, 16);
-
-        try {
-            $server = $srpService->getSRPServer();
-            $result = $server->step2(
-                $user->getEmail(),
-                $pf->salt,
-                $pf->verifier,
-                $A,
-                $B,
-                $b,
-                $M1
-            );
-
-            return [$result, $loginPayload, $userSecret];
-        } catch (InvalidSessionProofException) {
-            ErrorCode::INVALID_CREDENTIALS->throw();
-        }
     }
 }
